Add newly BluetoothHidDevice class for hid device role.
